Use go-junit-report for unit tests
