+++ magefile.go
@@ -15,7 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//+build mage
+//go:build mage
+// +build mage
 
 package main
 
@@ -189,9 +190,10 @@ func Test() error {
 				tst.UseBinaryIf(bin, useBinary),
 				tst.WithCoverage(path.Join(home, "cover.out")),
 				tst.Short(envTestShort),
-// 				tst.Out(bin), - due to bin file tests are failing on Windows, since it seems not to be used - it's disabled
+				// 				tst.Out(bin), - due to bin file tests are failing on Windows, since it seems not to be used - it's disabled
 				tst.Package(pkg),
 				tst.Verbose(true),
+				clitool.Flag("-timeout", "30m"),
 			)
 		})
 	})
