+++ pq/pq_rand_test.go
@@ -179,6 +179,7 @@ func runScenario(t *mint.T, cfg config, ops scenario) bool {
 
 		if op.exec != nil {
 			op.exec(ctx, qu)
+
 			if t.Failed() {
 				return false
 			}
